+++ contracts/src/index.ts
@@ -1,3 +1,4 @@
 export * from './font-manager';
 export * from './context';
 export * from './overlay';
+export * from './transmuter';