+++ contracts/src/transmuter.ts
@@ -1,0 +1,12 @@
+export type TransmuterOptions = {
+    config?: string;
+    theme?: string;
+};
+
+export interface Transmuter<
+    Input = string,
+    Output = unknown,
+    Options extends TransmuterOptions = TransmuterOptions
+> {
+    transmute(input: Input, options?: Options): Output;
+}